+++ autorest/codegen/models/list_type.py
@@ -28,13 +28,11 @@ class ListType(BaseType):
     def serialization_type(self) -> str:
         return f"[{self.element_type.serialization_type}]"
 
-    def type_annotation(self, *, is_operation_file: bool = False) -> str:
+    def type_annotation(self, **kwargs: Any) -> str:
         if self.code_model.options["version_tolerant"] and self.element_type.is_xml:
             # this means we're version tolerant XML, we just return the XML element
-            return self.element_type.type_annotation(
-                is_operation_file=is_operation_file
-            )
-        return f"List[{self.element_type.type_annotation(is_operation_file=is_operation_file)}]"
+            return self.element_type.type_annotation(**kwargs)
+        return f"List[{self.element_type.type_annotation(**kwargs)}]"
 
     def description(self, *, is_operation_file: bool) -> str:
         return "" if is_operation_file else self.yaml_data.get("description", "")
@@ -61,25 +59,23 @@ class ListType(BaseType):
 
         return ", ".join(attrs_list)
 
-    @property
-    def docstring_type(self) -> str:
+    def docstring_type(self, **kwargs: Any) -> str:
         if (
             self.code_model.options["version_tolerant"]
             and self.element_type.xml_metadata
         ):
             # this means we're version tolerant XML, we just return the XML element
-            return self.element_type.docstring_type
-        return f"list[{self.element_type.docstring_type}]"
+            return self.element_type.docstring_type(**kwargs)
+        return f"list[{self.element_type.docstring_type(**kwargs)}]"
 
-    @property
-    def docstring_text(self) -> str:
+    def docstring_text(self, **kwargs: Any) -> str:
         if (
             self.code_model.options["version_tolerant"]
             and self.element_type.xml_metadata
         ):
             # this means we're version tolerant XML, we just return the XML element
-            return self.element_type.docstring_text
-        return f"list of {self.element_type.docstring_text}"
+            return self.element_type.docstring_text(**kwargs)
+        return f"list of {self.element_type.docstring_text(**kwargs)}"
 
     @property
     def validation(self) -> Optional[Dict[str, Union[bool, int, str]]]:
@@ -126,7 +122,7 @@ class ListType(BaseType):
             ),
         )
 
-    def imports(self, *, is_operation_file: bool) -> FileImport:
+    def imports(self, **kwargs: Any) -> FileImport:
         file_import = FileImport()
         if not (
             self.code_model.options["version_tolerant"] and self.element_type.is_xml
@@ -134,7 +130,5 @@ class ListType(BaseType):
             file_import.add_submodule_import(
                 "typing", "List", ImportType.STDLIB, TypingSection.CONDITIONAL
             )
-        file_import.merge(
-            self.element_type.imports(is_operation_file=is_operation_file)
-        )
+        file_import.merge(self.element_type.imports(**kwargs))
         return file_import