+++ autorest/codegen/models/operation.py
@@ -91,7 +91,7 @@ class Operation(BaseBuilder[ParameterList]):  # pylint: disable=too-many-public-
             return "bool"
         response_body_annotations: OrderedSet[str] = {}
         for response in [r for r in self.responses if r.type]:
-            response_body_annotations[response.type_annotation()] = None
+            response_body_annotations[response.type_annotation(**kwargs)] = None
         response_str = ", ".join(response_body_annotations.keys()) or "None"
         if len(response_body_annotations) > 1:
             return f"Union[{response_str}]"
@@ -107,7 +107,7 @@ class Operation(BaseBuilder[ParameterList]):  # pylint: disable=too-many-public-
             return "ClsType[None]"
         return f"ClsType[{self.response_type_annotation(async_mode=async_mode)}]"
 
-    def _response_docstring_helper(self, attr_name: str) -> str:
+    def _response_docstring_helper(self, attr_name: str, **kwargs: Any) -> str:
         responses_with_body = [r for r in self.responses if r.type]
         if (
             self.request_builder.method.lower() == "head"
@@ -116,7 +116,8 @@ class Operation(BaseBuilder[ParameterList]):  # pylint: disable=too-many-public-
             return "bool"
         if responses_with_body:
             response_docstring_values: OrderedSet[str] = {
-                getattr(response, attr_name): None for response in responses_with_body
+                getattr(response, attr_name)(**kwargs): None
+                for response in responses_with_body
             }
             retval = " or ".join(response_docstring_values.keys())
             if self.has_optional_return_type:
@@ -125,13 +126,13 @@ class Operation(BaseBuilder[ParameterList]):  # pylint: disable=too-many-public-
         return "None"
 
     def response_docstring_text(self, **kwargs) -> str:
-        retval = self._response_docstring_helper("docstring_text")
+        retval = self._response_docstring_helper("docstring_text", **kwargs)
         if not self.code_model.options["version_tolerant"]:
             retval += " or the result of cls(response)"
         return retval
 
     def response_docstring_type(self, **kwargs) -> str:
-        return self._response_docstring_helper("docstring_type")
+        return self._response_docstring_helper("docstring_type", **kwargs)
 
     @property
     def has_response_body(self) -> bool:
@@ -168,30 +169,28 @@ class Operation(BaseBuilder[ParameterList]):  # pylint: disable=too-many-public-
             )
         )
 
-    def _imports_shared(
-        self, async_mode: bool  # pylint: disable=unused-argument
-    ) -> FileImport:
+    def _imports_shared(self, async_mode: bool) -> FileImport:
         file_import = FileImport()
         file_import.add_submodule_import(
             "typing", "Any", ImportType.STDLIB, TypingSection.CONDITIONAL
         )
         if not self.abstract:
             for param in self.parameters.method:
-                file_import.merge(param.imports())
+                file_import.merge(param.imports(async_mode))
 
         for response in self.responses:
-            file_import.merge(response.imports())
+            file_import.merge(response.imports(async_mode=async_mode))
 
-        response_types = [r.type_annotation() for r in self.responses if r.type]
+        response_types = [
+            r.type_annotation(async_mode=async_mode) for r in self.responses if r.type
+        ]
         if len(set(response_types)) > 1:
             file_import.add_submodule_import(
                 "typing", "Union", ImportType.STDLIB, TypingSection.CONDITIONAL
             )
         return file_import
 
-    def imports_for_multiapi(
-        self, async_mode: bool
-    ) -> FileImport:  # pylint: disable=unused-argument
+    def imports_for_multiapi(self, async_mode: bool) -> FileImport:
         return self._imports_shared(async_mode)
 
     def imports(self, async_mode: bool, is_python3_file: bool) -> FileImport: