+++ autorest/codegen/serializers/client_serializer.py
@@ -22,7 +22,7 @@ class ClientSerializer:
             method_name="__init__",
             need_self_param=True,
             method_param_signatures=self.code_model.client.parameters.method_signature(
-                async_mode or self.is_python3_file
+                async_mode or self.is_python3_file, async_mode
             ),
         )
 
@@ -72,7 +72,7 @@ class ClientSerializer:
                 f":{param.description_keyword} {param.client_name}: {param.description}"
             )
             retval.append(
-                f":{param.docstring_type_keyword} {param.client_name}: {param.docstring_type}"
+                f":{param.docstring_type_keyword} {param.client_name}: {param.docstring_type(async_mode=async_mode)}"
             )
         if self.code_model.has_lro_operations:
             retval.append(
@@ -265,7 +265,7 @@ class ConfigSerializer:
             method_name="__init__",
             need_self_param=True,
             method_param_signatures=self.code_model.config.parameters.method_signature(
-                async_mode or self.is_python3_file
+                async_mode or self.is_python3_file, async_mode
             ),
         )
 
@@ -302,12 +302,12 @@ class ConfigSerializer:
             if not (p.optional or p.constant)
         ]
 
-    def property_descriptions(self) -> List[str]:
+    def property_descriptions(self, async_mode: bool) -> List[str]:
         retval: List[str] = []
         for p in self.code_model.config.parameters.method:
             retval.append(f":{p.description_keyword} {p.client_name}: {p.description}")
             retval.append(
-                f":{p.docstring_type_keyword} {p.client_name}: {p.docstring_type}"
+                f":{p.docstring_type_keyword} {p.client_name}: {p.docstring_type(async_mode=async_mode)}"
             )
         retval.append('"""')
         return retval